pyproject.toml
[tool.poetry]
name = "generativeai-apps"
version = "0.1.0"
description = ""
authors = ["y-hiroki-radiotech <bnndd439@ybb.ne.jp>"]
readme = "README.md"
packages = [{include = "generativeai_apps"}]

[tool.poetry.dependencies]
python = "3.11.3"
python-dotenv = "^1.0.1"
google-generativeai = "^0.8.3"
ipython = "^8.31.0"
streamlit = "^1.41.1"
pypdf2 = "^3.0.1"
langchain = "^0.3.13"
chromadb = "^0.5.23"
faiss-cpu = "^1.9.0.post1"
langchain-google-genai = "^2.0.7"
langchain-groq = "^0.2.2"
groq = "^0.13.1"
langchain-community = "^0.3.13"
pypdf = "^5.1.0"
pdf2image = "^1.17.0"
youtube-transcript-api = "^0.6.3"
crewai = "^0.86.0"
crewai-tools = "^0.17.0"
streamlit-extras = "^0.5.0"
google-genai = "^0.3.0"
phidata = "^2.7.6"
yfinance = "^0.2.51"
packaging = "^24.2"
duckduckgo-search = "^7.2.0"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
pyown = "^0.3.0"
unstructured = "^0.16.12"
langgraph = "0.2.22"
pydantic = "^2.10.4"
pydantic-core = "^2.27.2"
langgraph-checkpoint = "1.0.11"
argparse = "^1.4.0"
langchainhub = "^0.1.21"
bs4 = "^0.0.2"
tiktoken = "0.7"
huggingface-hub = "^0.27.1"
transformers = "<4.48.0"
sentencepiece = "^0.2.0"
ollama = "^0.4.7"
gradio = "^5.13.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
